Add compression function, change files names from grayscape.hpp/cpp to compresson.hpp/cpp
