+++ module1/homework/grayscaleImages/compression.hpp
@@ -5,5 +5,5 @@
 constexpr size_t width = 32;
 constexpr size_t height = 32;
 
-std::vector<std::pair<uint8_t, uint8_t>> compressGrayscale(std::array<std::array<uint8_t, width>, height>);
+std::vector<std::pair<uint8_t, uint8_t>> compressGrayscale(const std::array<std::array<uint8_t, width>, height>&);
 void printMap (const std::array<std::array<uint8_t, width>, height>&);
