+++ module1/homework/grayscaleImages/main.cpp
@@ -1,7 +1,7 @@
 #include <array>
 #include <forward_list>
 
-#include "grayscale.hpp"
+#include "compression.hpp"
 
 
 std::array<std::array<uint8_t, 32>, 32> generateNinja() {
@@ -43,7 +43,7 @@ std::array<std::array<uint8_t, 32>, 32> generateNinja() {
 
 int main() {
     auto ninja = generateNinja();
-    printMap(ninja);
+    // printMap(ninja);
     auto compressed = compressGrayscale(ninja);
     auto decompressed = decompressGrayscale(compressed);
     // printMap(decompressed);
